producer/requirements.txt
watchdog
kafka-python
csv-diff
pyyaml
